Update ScrollView Content update methods to handle null content Fixes #2367
