+++ src/Core/src/Handlers/ScrollView/ScrollViewHandler.Android.cs
@@ -38,10 +38,10 @@ namespace Microsoft.Maui.Handlers
 
 		public static void MapContent(ScrollViewHandler handler, IScrollView scrollView)
 		{
-			if (handler.NativeView == null || handler.MauiContext == null || scrollView.PresentedContent == null)
+			if (handler.NativeView == null || handler.MauiContext == null)
 				return;
 
-			handler.NativeView.SetContent(scrollView.PresentedContent.ToNative(handler.MauiContext));
+			handler.NativeView.UpdateContent(scrollView.PresentedContent, handler.MauiContext);
 		}
 
 		public static void MapHorizontalScrollBarVisibility(ScrollViewHandler handler, IScrollView scrollView)
