+++ src/Core/src/Handlers/ScrollView/ScrollViewHandler.Windows.cs
@@ -28,10 +28,10 @@ namespace Microsoft.Maui.Handlers
 
 		public static void MapContent(ScrollViewHandler handler, IScrollView scrollView)
 		{
-			if (handler.NativeView == null || handler.MauiContext == null || scrollView.Content == null)
+			if (handler.NativeView == null || handler.MauiContext == null)
 				return;
 
-			handler.NativeView.Content = scrollView.PresentedContent?.ToNative(handler.MauiContext);
+			handler.NativeView.UpdateContent(scrollView.PresentedContent, handler.MauiContext);
 		}
 
 		public static void MapHorizontalScrollBarVisibility(ScrollViewHandler handler, IScrollView scrollView)
