+++ src/Core/src/Handlers/ScrollView/ScrollViewHandler.iOS.cs
@@ -67,10 +67,10 @@ namespace Microsoft.Maui.Handlers
 
 		public static void MapContent(ScrollViewHandler handler, IScrollView scrollView)
 		{
-			if (handler.NativeView == null || handler.MauiContext == null || scrollView.PresentedContent == null)
+			if (handler.NativeView == null || handler.MauiContext == null)
 				return;
 
-			handler.NativeView.UpdateContent(scrollView.PresentedContent.ToNative(handler.MauiContext));
+			handler.NativeView.UpdateContent(scrollView.PresentedContent, handler.MauiContext);
 		}
 
 		public static void MapContentSize(ScrollViewHandler handler, IScrollView scrollView)
