+++ src/Core/src/Platform/Android/MauiScrollView.cs
@@ -376,34 +376,4 @@ namespace Microsoft.Maui
 		bool ScrollbarFadingEnabled { get; set; }
 		void AwakenScrollBars();
 	}
-
-	internal static class ScrollViewExtensions
-	{
-		internal static void HandleScrollBarVisibilityChange(this IScrollBarView scrollView)
-		{
-			// According to the Android Documentation
-			// * <p>AwakenScrollBars method should be invoked every time a subclass directly updates
-			// *the scroll parameters.</ p >
-
-			// If AwakenScrollBars is never called there are cases where the ScrollDrawable is never called
-			// which causes a crash during draw
-
-			if (scrollView.ScrollBarsInitialized)
-				scrollView.AwakenScrollBars();
-
-			// The scrollbar drawable won't initialize if ScrollbarFadingEnabled == false
-			if (!scrollView.ScrollbarFadingEnabled)
-			{
-				scrollView.ScrollbarFadingEnabled = true;
-				scrollView.AwakenScrollBars();
-				scrollView.ScrollbarFadingEnabled = false;
-			}
-			else
-			{
-				scrollView.AwakenScrollBars();
-			}
-
-			scrollView.ScrollBarsInitialized = true;
-		}
-	}
 }
