+++ src/Core/src/Platform/Windows/ScrollViewerExtensions.cs
@@ -49,5 +49,10 @@ namespace Microsoft.Maui
 			// Note that the Orientation setting of "Neither" is covered by the measurement code (the size of the content is limited
 			// so that no scrolling is possible) and the xplat scrolling code (the ScrollTo methods are disabled when Orientation=Neither)
 		}
+
+		public static void UpdateContent(this ScrollViewer scrollViewer, IView? content, IMauiContext context)
+		{
+			scrollViewer.Content = content == null ? null : content.ToNative(context);
+		}
 	}
 }
